requirements.txt
streamlit
torch
torchvision
pillow
pandas
openpyxl
requests
